fix: migrate add-link to zustand
